backend/pyproject.toml
[project]
name = "administracion-conjunta-api"
version = "0.1.0"
description = "Backend Python para Administracion Conjunta"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.116.0,<1.0.0",
  "uvicorn[standard]>=0.35.0,<1.0.0",
  "pydantic-settings>=2.10.0,<3.0.0",
  "firebase-admin>=7.1.0,<8.0.0",
  "python-multipart>=0.0.20,<1.0.0",
  "openpyxl>=3.1.0,<4.0.0",
  "fpdf2>=2.7.0,<3.0.0",
]

[project.optional-dependencies]
dev = [
  "httpx>=0.28.0,<1.0.0",
  "pytest>=8.4.0,<9.0.0",
  "ruff>=0.12.0,<1.0.0",
]
performance = [
  "locust>=2.37.0,<3.0.0",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"
